Fix some compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/3222



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@14114 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/cintltst/nfsprep.c
@@ -266,7 +266,7 @@ nfs4_mixed_prepare( const char* src, int32_t srcLength,
         return 0;
     }
     if(srcLength == -1){
-        srcLength = strlen(src);
+        srcLength = (int32_t)strlen(src);
     }
     getPrefixSuffix(src, srcLength, &prefix, &prefixLen, &suffix, &suffixLen, status);
